requirements.txt
anyio==3.6.1
APScheduler==3.9.1
async-generator==1.10
attrs==21.4.0
autopep8==1.7.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
chromedriver-autoinstaller==0.3.1
click==8.1.3
cryptography==37.0.4
fake-useragent==0.1.11
Flask==2.2.2
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
outcome==1.2.0
pycodestyle==2.9.1
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-telegram-bot==20.0a4
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rfc3986==1.5.0
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
toml==0.10.2
tornado==6.2
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.10
Werkzeug==2.2.2
wsproto==1.1.0
zipp==3.8.1
